requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0

# RAG功能依赖
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0

# 注：本地LLM服务已移至 WSL vLLM，Windows 侧通过 HTTP 调用
# vLLM 安装请参考 ~/llm-server/requirements.txt (WSL侧)
